Modified ProductController to validate if the product exists before deleting it. Created Middleware AdminLoggedIn. Registered AdminLoggedIn middleware as 'admin' inside app/Http/kernel.js. Added admin middleware to products.store, products.update and products.destroy routes. Created migration to alter email as a unique column for users table.

+++ app/Http/Controllers/api/ProductController.js
@@ -66,8 +66,11 @@ class ProductController {
   * destroy(request, response) {
     const productId = request.param('id')
     const product = yield Product.find(productId)
-    yield product.delete()
-    response.ok('Product was deleted')
+    if (product) {
+      yield product.delete()
+      response.ok('Product was deleted')
+    }
+    response.notFound('Product was not found')
   }
 
 }
